requirements.txt
google-genai
streamlit
pandas
requests